chore: create state for balance combined and optimize applyBalanceDiffs function
