+++ apps/cowswap-frontend/src/modules/combinedBalances/state/balanceCombinedAtom.ts
@@ -1,0 +1,5 @@
+import { atomWithReset } from 'jotai/utils'
+
+import { BalancesState } from '@cowprotocol/balances-and-allowances'
+
+export const balancesCombinedAtom = atomWithReset<BalancesState>({ isLoading: false, values: {} })
